FileManager now handles the file->quit menu item
We may need to save the model before quitting. Keeping it in MainWindow ment that MainWindow is having a dependency on FileManager. Since FileManager also has a dependency on MainWindow that would mean we'll end up with a cyclic dependency. Now the cycle is broken.
